Instead of having two test classes, for essentially the exact same thing, I've deleted the Api\GetInstitutesTest and Api\ListInstitutionsBase classes.
